requirements.txt
pylint==2.7.4
fastapi==0.78.0
uvicorn==0.17.6
black==22.3.0
pyhumps==3.7.1
sqlalchemy==1.4.37
